requirements.txt
tensorflow~=2.4.1
matplotlib~=3.3.3
numpy~=1.19.5
Pillow~=8.0.1
scipy~=1.5.4
scikit-learn~=0.23.2